aider: Refactored interpolation method to use `numpy.interp` for smoother transitions in `make_density_video`.
# Aider chat conversation:

USER: please make interpolate_data more smoothly

ASSISTANT: To make the interpolation of data smoother in the `make_density_video` function within the `twitter/visualisation.py` file, we can use a more sophisticated interpolation method. One common approach is to use a spline interpolation or a similar smooth interpolation technique.

However, since the current code does not use any specific interpolation functions and simply fills the array based on the progress, we will need to introduce a new interpolation method. We can use `numpy.interp` for linear interpolation or `scipy.interpolate` for more advanced spline interpolation.

For this example, I will show you how to use `numpy.interp` for linear interpolation to achieve smoother transitions. If you need a more advanced interpolation, please let me know, and we can use `scipy.interpolate` instead.

Here is the *SEARCH/REPLACE* block to modify the interpolation in `twitter/visualisation.py`:

twitter/visualisation.py
```python
<<<<<<< SEARCH
    # Interpolate data for the graph that changes from left to right
    for i in range(1, num_frames):
        progress = i / num_frames
        interpolated_data[i, :int(progress * len(data))] = data[:int(progress * len(data))]
=======
    # Generate time points for original and interpolated data
    original_time_points = np.linspace(0, 1, len(data))
    interpolated_time_points = np.linspace(0, 1, num_frames)

    # Interpolate data for the graph that changes from left to right
    for i, t in enumerate(interpolated_time_points):
        interpolated_data[i] = np.interp(t, original_time_points, data)
>>>>>>> REPLACE
```

Please note that this change assumes that the `data` array is one-dimensional and that the interpolation is done over the entire range of `data`. If `data` is not one-dimensional or if you have other requirements for the interpolation, additional adjustments may be necessary.